backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-login
pymysql
cryptography
mariadb
gspread
oauth2client
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
requests
